+++ test/TestAsset/ConnectionMandatoryRecursiveContainerIdConfiguration.php
@@ -10,14 +10,12 @@
 namespace InteropTest\Config\TestAsset;
 
 use Interop\Config\ConfigurationTrait;
-use Interop\Config\HasContainerId;
-use Interop\Config\HasMandatoryOptions;
-use Interop\Config\ObtainsOptions;
+use Interop\Config\RequiresContainerId;
+use Interop\Config\RequiresMandatoryOptions;
 
 class ConnectionMandatoryRecursiveContainerIdConfiguration implements
-    ObtainsOptions,
-    HasContainerId,
-    HasMandatoryOptions
+    RequiresContainerId,
+    RequiresMandatoryOptions
 {
     use ConfigurationTrait;